+++ src/game/camera_view.cpp
@@ -1,14 +1,14 @@
 #include "camera_view.h"}
 
 CameraView::CameraView(const CameraView& other) {
-    for (int row = 0; row < matrixSize; ++row) {
+    for (byte row = 0; row < matrixSize; ++row) {
         this->mapMatrix[row] = other.getMapRow(row);
         this->coinsMatrix[row] = other.getCoinsRow(row);
     }
 }
 
 CameraView CameraView::operator = (const CameraView& other) {
-    for (int row = 0; row < matrixSize; ++row)  {
+    for (byte row = 0; row < matrixSize; ++row)  {
         this->mapMatrix[row] = other.getMapRow(row);
         this->coinsMatrix[row] = other.getCoinsRow(row);
     }
@@ -27,7 +27,7 @@ void CameraView::setPosition(Point p, bool value) {
 }
 
 void CameraView::changeMap(byte* newMapColumns, byte* newCoinsColumns) {
-    for (int col = 0; col < matrixSize; ++col) {
+    for (byte col = 0; col < matrixSize; ++col) {
         mapColumns[col] = newMapColumns[col];
         coinsColumns[col] = newCoinsColumns[col];
     }
@@ -35,17 +35,17 @@ void CameraView::changeMap(byte* newMapColumns, byte* newCoinsColumns) {
 }
 
 void CameraView::updateMap() {
-    for (int col = 0; col < matrixSize; ++col) {
+    for (byte col = 0; col < matrixSize; ++col) {
         setColumn(mapMatrix, matrixSize - col - 1, mapColumns[col]);      // complementar position due to matrix logic
         setColumn(coinsMatrix, matrixSize - col - 1, coinsColumns[col]);  // complementar position due to matrix logic
     }
 }
 
-byte CameraView::getColumn(byte* bytesMatrix, int index) {
+byte CameraView::getColumn(byte* bytesMatrix, byte index) {
     byte mask = 1 << index;
     byte answer = 0;
 
-    for (int i = 0; i < matrixSize; ++i) {
+    for (byte i = 0; i < matrixSize; ++i) {
         byte byteValue = (bytesMatrix[i] & mask);
         answer <<= 1;
         if (byteValue != 0)
@@ -55,11 +55,11 @@ byte CameraView::getColumn(byte* bytesMatrix, int index) {
     return answer;
 }
 
-void CameraView::setColumn(byte* bytesMatrix, int index, byte value) {
+void CameraView::setColumn(byte* bytesMatrix, byte index, byte value) {
     byte setMask = 1 << index;                                  // zero everywhere exept on index
     byte unsetMask = ((1 << matrixSize) - 1) ^ (1 << index);    // one everywhere exept on index
 
-    for (int i = 0; i < matrixSize; ++i) {
+    for (byte i = 0; i < matrixSize; ++i) {
         byte byteValue = value & (1 << (matrixSize - i - 1));
         if (byteValue != 0)
             bytesMatrix[i] |= setMask;
@@ -69,7 +69,7 @@ void CameraView::setColumn(byte* bytesMatrix, int index, byte value) {
 }
 
 void CameraView::appendColumn(byte mapValue, byte coinsValue) {
-    for (int col = matrixSize - 1; col > 0; --col) {
+    for (byte col = matrixSize - 1; col > 0; --col) {
         setColumn(mapMatrix, col, getColumn(mapMatrix, col - 1));
         setColumn(coinsMatrix, col, getColumn(coinsMatrix, col - 1));
     }
@@ -79,7 +79,7 @@ void CameraView::appendColumn(byte mapValue, byte coinsValue) {
 }
 
 void CameraView::prependColumn(byte mapValue, byte coinsValue) {
-    for (int col = 0; col < matrixSize - 1; ++col) {
+    for (byte col = 0; col < matrixSize - 1; ++col) {
         setColumn(mapMatrix, col, getColumn(mapMatrix, col + 1));
         setColumn(coinsMatrix, col, getColumn(coinsMatrix, col + 1));
     }
@@ -89,7 +89,7 @@ void CameraView::prependColumn(byte mapValue, byte coinsValue) {
 }
 
 void CameraView::changeCoinsState(bool state) {
-    for (int row = 0; row < matrixSize; ++row) {
+    for (byte row = 0; row < matrixSize; ++row) {
         if (state == true) 
             mapMatrix[row] |= coinsMatrix[row];
         else
